feat: split schemas
