+++ packages/astro/src/core/config/index.ts
@@ -6,6 +6,6 @@ export {
 } from './config.js';
 export { createNodeLogger } from './logging.js';
 export { mergeConfig } from './merge.js';
-export type { AstroConfigType } from './schema.js';
+export type { AstroConfigType } from './schemas/index.js';
 export { createSettings } from './settings.js';
 export { loadTSConfig, updateTSConfigForFramework } from './tsconfig.js';
