+++ packages/astro/src/core/config/schemas/base.ts
@@ -1,6 +1,4 @@
 import type { OutgoingHttpHeaders } from 'node:http';
-import path from 'node:path';
-import { fileURLToPath, pathToFileURL } from 'node:url';
 import type {
 	ShikiConfig,
 	RehypePlugin as _RehypePlugin,
@@ -10,10 +8,9 @@ import type {
 import { markdownConfigDefaults, syntaxHighlightDefaults } from '@astrojs/markdown-remark';
 import { type BuiltinTheme, bundledThemes } from 'shiki';
 import { z } from 'zod';
-import type { SvgRenderMode } from '../../assets/utils/svg.js';
-import { EnvSchema } from '../../env/schema.js';
-import type { AstroConfig, AstroUserConfig, ViteUserConfig } from '../../types/public/config.js';
-import { appendForwardSlash, prependForwardSlash, removeTrailingForwardSlash } from '../path.js';
+import type { SvgRenderMode } from '../../../assets/utils/svg.js';
+import { EnvSchema } from '../../../env/schema.js';
+import type { AstroUserConfig, ViteUserConfig } from '../../../types/public/config.js';
 
 // The below types are required boilerplate to workaround a Zod issue since v3.21.2. Since that version,
 // Zod's compiled TypeScript would "simplify" certain values to their base representation, causing references
@@ -35,7 +32,7 @@ import { appendForwardSlash, prependForwardSlash, removeTrailingForwardSlash } f
 // back to the issue again. The complexified type should be the base representation that we want to expose.
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-interface ComplexifyUnionObj {}
+export interface ComplexifyUnionObj {}
 
 type ComplexifyWithUnion<T> = T & ComplexifyUnionObj;
 type ComplexifyWithOmit<T> = Omit<T, '__nonExistent'>;
@@ -45,7 +42,7 @@ type ShikiTheme = ComplexifyWithUnion<NonNullable<ShikiConfig['theme']>>;
 type ShikiTransformer = ComplexifyWithUnion<NonNullable<ShikiConfig['transformers']>[number]>;
 type RehypePlugin = ComplexifyWithUnion<_RehypePlugin>;
 type RemarkPlugin = ComplexifyWithUnion<_RemarkPlugin>;
-type RemarkRehype = ComplexifyWithOmit<_RemarkRehype>;
+export type RemarkRehype = ComplexifyWithOmit<_RemarkRehype>;
 
 export const ASTRO_CONFIG_DEFAULTS = {
 	root: '.',
@@ -521,336 +518,3 @@ export const AstroConfigSchema = z.object({
 });
 
 export type AstroConfigType = z.infer<typeof AstroConfigSchema>;
-
-export function createRelativeSchema(cmd: string, fileProtocolRoot: string) {
-	let originalBuildClient: string;
-	let originalBuildServer: string;
-
-	// We need to extend the global schema to add transforms that are relative to root.
-	// This is type checked against the global schema to make sure we still match.
-	const AstroConfigRelativeSchema = AstroConfigSchema.extend({
-		root: z
-			.string()
-			.default(ASTRO_CONFIG_DEFAULTS.root)
-			.transform((val) => resolveDirAsUrl(val, fileProtocolRoot)),
-		srcDir: z
-			.string()
-			.default(ASTRO_CONFIG_DEFAULTS.srcDir)
-			.transform((val) => resolveDirAsUrl(val, fileProtocolRoot)),
-		compressHTML: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.compressHTML),
-		publicDir: z
-			.string()
-			.default(ASTRO_CONFIG_DEFAULTS.publicDir)
-			.transform((val) => resolveDirAsUrl(val, fileProtocolRoot)),
-		outDir: z
-			.string()
-			.default(ASTRO_CONFIG_DEFAULTS.outDir)
-			.transform((val) => resolveDirAsUrl(val, fileProtocolRoot)),
-		cacheDir: z
-			.string()
-			.default(ASTRO_CONFIG_DEFAULTS.cacheDir)
-			.transform((val) => resolveDirAsUrl(val, fileProtocolRoot)),
-		build: z
-			.object({
-				format: z
-					.union([z.literal('file'), z.literal('directory'), z.literal('preserve')])
-					.optional()
-					.default(ASTRO_CONFIG_DEFAULTS.build.format),
-				// NOTE: `client` and `server` are transformed relative to the default outDir first,
-				// later we'll fix this to be relative to the actual `outDir`
-				client: z
-					.string()
-					.optional()
-					.default(ASTRO_CONFIG_DEFAULTS.build.client)
-					.transform((val) => {
-						originalBuildClient = val;
-						return resolveDirAsUrl(
-							val,
-							path.resolve(fileProtocolRoot, ASTRO_CONFIG_DEFAULTS.outDir),
-						);
-					}),
-				server: z
-					.string()
-					.optional()
-					.default(ASTRO_CONFIG_DEFAULTS.build.server)
-					.transform((val) => {
-						originalBuildServer = val;
-						return resolveDirAsUrl(
-							val,
-							path.resolve(fileProtocolRoot, ASTRO_CONFIG_DEFAULTS.outDir),
-						);
-					}),
-				assets: z.string().optional().default(ASTRO_CONFIG_DEFAULTS.build.assets),
-				assetsPrefix: z
-					.string()
-					.optional()
-					.or(z.object({ fallback: z.string() }).and(z.record(z.string())).optional()),
-				serverEntry: z.string().optional().default(ASTRO_CONFIG_DEFAULTS.build.serverEntry),
-				redirects: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.build.redirects),
-				inlineStylesheets: z
-					.enum(['always', 'auto', 'never'])
-					.optional()
-					.default(ASTRO_CONFIG_DEFAULTS.build.inlineStylesheets),
-				concurrency: z.number().min(1).optional().default(ASTRO_CONFIG_DEFAULTS.build.concurrency),
-			})
-			.optional()
-			.default({}),
-		server: z.preprocess(
-			// preprocess
-			(val) => {
-				if (typeof val === 'function') {
-					return val({ command: cmd === 'dev' ? 'dev' : 'preview' });
-				} else {
-					return val;
-				}
-			},
-			// validate
-			z
-				.object({
-					open: z
-						.union([z.string(), z.boolean()])
-						.optional()
-						.default(ASTRO_CONFIG_DEFAULTS.server.open),
-					host: z
-						.union([z.string(), z.boolean()])
-						.optional()
-						.default(ASTRO_CONFIG_DEFAULTS.server.host),
-					port: z.number().optional().default(ASTRO_CONFIG_DEFAULTS.server.port),
-					headers: z.custom<OutgoingHttpHeaders>().optional(),
-					streaming: z.boolean().optional().default(true),
-					allowedHosts: z
-						.union([z.array(z.string()), z.literal(true)])
-						.optional()
-						.default(ASTRO_CONFIG_DEFAULTS.server.allowedHosts),
-				})
-				.optional()
-				.default({}),
-		),
-	}).transform((config) => {
-		// If the user changed `outDir`, we need to also update `build.client` and `build.server`
-		// the be based on the correct `outDir`
-		if (
-			config.outDir.toString() !==
-			resolveDirAsUrl(ASTRO_CONFIG_DEFAULTS.outDir, fileProtocolRoot).toString()
-		) {
-			const outDirPath = fileURLToPath(config.outDir);
-			config.build.client = resolveDirAsUrl(originalBuildClient, outDirPath);
-			config.build.server = resolveDirAsUrl(originalBuildServer, outDirPath);
-		}
-
-		// Handle `base` and `image.endpoint.route` trailing slash based on `trailingSlash` config
-		if (config.trailingSlash === 'never') {
-			config.base = prependForwardSlash(removeTrailingForwardSlash(config.base));
-			config.image.endpoint.route = prependForwardSlash(
-				removeTrailingForwardSlash(config.image.endpoint.route),
-			);
-		} else if (config.trailingSlash === 'always') {
-			config.base = prependForwardSlash(appendForwardSlash(config.base));
-			config.image.endpoint.route = prependForwardSlash(
-				appendForwardSlash(config.image.endpoint.route),
-			);
-		} else {
-			config.base = prependForwardSlash(config.base);
-			config.image.endpoint.route = prependForwardSlash(config.image.endpoint.route);
-		}
-
-		return config;
-	});
-
-	return AstroConfigRelativeSchema;
-}
-
-export const AstroConfigRefinedSchema = z
-	.custom<AstroConfig>()
-	.superRefine((config, ctx) => {
-		if (
-			config.build.assetsPrefix &&
-			typeof config.build.assetsPrefix !== 'string' &&
-			!config.build.assetsPrefix.fallback
-		) {
-			ctx.addIssue({
-				code: z.ZodIssueCode.custom,
-				message: 'The `fallback` is mandatory when defining the option as an object.',
-				path: ['build', 'assetsPrefix'],
-			});
-		}
-
-		for (let i = 0; i < config.image.remotePatterns.length; i++) {
-			const { hostname, pathname } = config.image.remotePatterns[i];
-
-			if (
-				hostname &&
-				hostname.includes('*') &&
-				!(hostname.startsWith('*.') || hostname.startsWith('**.'))
-			) {
-				ctx.addIssue({
-					code: z.ZodIssueCode.custom,
-					message: 'wildcards can only be placed at the beginning of the hostname',
-					path: ['image', 'remotePatterns', i, 'hostname'],
-				});
-			}
-
-			if (
-				pathname &&
-				pathname.includes('*') &&
-				!(pathname.endsWith('/*') || pathname.endsWith('/**'))
-			) {
-				ctx.addIssue({
-					code: z.ZodIssueCode.custom,
-					message: 'wildcards can only be placed at the end of a pathname',
-					path: ['image', 'remotePatterns', i, 'pathname'],
-				});
-			}
-		}
-
-		if (
-			config.i18n &&
-			typeof config.i18n.routing !== 'string' &&
-			!config.i18n.routing.redirectToDefaultLocale &&
-			!config.i18n.routing.prefixDefaultLocale
-		) {
-			ctx.addIssue({
-				code: z.ZodIssueCode.custom,
-				message:
-					'The option `i18n.routing.redirectToDefaultLocale` is only useful when the `i18n.routing.prefixDefaultLocale` is set to `true`. Remove the option `i18n.routing.redirectToDefaultLocale`, or change its value to `true`.',
-				path: ['i18n', 'routing', 'redirectToDefaultLocale'],
-			});
-		}
-
-		if (config.outDir.toString().startsWith(config.publicDir.toString())) {
-			ctx.addIssue({
-				code: z.ZodIssueCode.custom,
-				message:
-					'The value of `outDir` must not point to a path within the folder set as `publicDir`, this will cause an infinite loop',
-				path: ['outDir'],
-			});
-		}
-
-		if (config.i18n) {
-			const { defaultLocale, locales: _locales, fallback, domains } = config.i18n;
-			const locales = _locales.map((locale) => {
-				if (typeof locale === 'string') {
-					return locale;
-				} else {
-					return locale.path;
-				}
-			});
-			if (!locales.includes(defaultLocale)) {
-				ctx.addIssue({
-					code: z.ZodIssueCode.custom,
-					message: `The default locale \`${defaultLocale}\` is not present in the \`i18n.locales\` array.`,
-					path: ['i18n', 'locales'],
-				});
-			}
-			if (fallback) {
-				for (const [fallbackFrom, fallbackTo] of Object.entries(fallback)) {
-					if (!locales.includes(fallbackFrom)) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message: `The locale \`${fallbackFrom}\` key in the \`i18n.fallback\` record doesn't exist in the \`i18n.locales\` array.`,
-							path: ['i18n', 'fallbacks'],
-						});
-					}
-
-					if (fallbackFrom === defaultLocale) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message: `You can't use the default locale as a key. The default locale can only be used as value.`,
-							path: ['i18n', 'fallbacks'],
-						});
-					}
-
-					if (!locales.includes(fallbackTo)) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message: `The locale \`${fallbackTo}\` value in the \`i18n.fallback\` record doesn't exist in the \`i18n.locales\` array.`,
-							path: ['i18n', 'fallbacks'],
-						});
-					}
-				}
-			}
-			if (domains) {
-				const entries = Object.entries(domains);
-				const hasDomains = domains ? Object.keys(domains).length > 0 : false;
-				if (entries.length > 0 && !hasDomains) {
-					ctx.addIssue({
-						code: z.ZodIssueCode.custom,
-						message: `When specifying some domains, the property \`i18n.routing.strategy\` must be set to \`"domains"\`.`,
-						path: ['i18n', 'routing', 'strategy'],
-					});
-				}
-
-				if (hasDomains) {
-					if (!config.site) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message:
-								"The option `site` isn't set. When using the 'domains' strategy for `i18n`, `site` is required to create absolute URLs for locales that aren't mapped to a domain.",
-							path: ['site'],
-						});
-					}
-					if (config.output !== 'server') {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message: 'Domain support is only available when `output` is `"server"`.',
-							path: ['output'],
-						});
-					}
-				}
-
-				for (const [domainKey, domainValue] of entries) {
-					if (!locales.includes(domainKey)) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message: `The locale \`${domainKey}\` key in the \`i18n.domains\` record doesn't exist in the \`i18n.locales\` array.`,
-							path: ['i18n', 'domains'],
-						});
-					}
-					if (!domainValue.startsWith('https') && !domainValue.startsWith('http')) {
-						ctx.addIssue({
-							code: z.ZodIssueCode.custom,
-							message:
-								"The domain value must be a valid URL, and it has to start with 'https' or 'http'.",
-							path: ['i18n', 'domains'],
-						});
-					} else {
-						try {
-							const domainUrl = new URL(domainValue);
-							if (domainUrl.pathname !== '/') {
-								ctx.addIssue({
-									code: z.ZodIssueCode.custom,
-									message: `The URL \`${domainValue}\` must contain only the origin. A subsequent pathname isn't allowed here. Remove \`${domainUrl.pathname}\`.`,
-									path: ['i18n', 'domains'],
-								});
-							}
-						} catch {
-							// no need to catch the error
-						}
-					}
-				}
-			}
-		}
-
-		if (
-			!config.experimental.responsiveImages &&
-			(config.image.experimentalLayout ||
-				config.image.experimentalObjectFit ||
-				config.image.experimentalObjectPosition ||
-				config.image.experimentalBreakpoints)
-		) {
-			ctx.addIssue({
-				code: z.ZodIssueCode.custom,
-				message:
-					'The `experimentalLayout`, `experimentalObjectFit`, `experimentalObjectPosition` and `experimentalBreakpoints` options are only available when `experimental.responsiveImages` is enabled.',
-				path: ['experimental', 'responsiveImages'],
-			});
-		}
-	});
-
-function resolveDirAsUrl(dir: string, root: string) {
-	let resolvedDir = path.resolve(root, dir);
-	if (!resolvedDir.endsWith(path.sep)) {
-		resolvedDir += path.sep;
-	}
-	return pathToFileURL(resolvedDir);
-}
