+++ packages/astro/src/core/config/schemas/index.ts
@@ -1,0 +1,3 @@
+export { AstroConfigSchema, ASTRO_CONFIG_DEFAULTS, type AstroConfigType } from './base.js';
+export { createRelativeSchema } from './relative.js';
+export { AstroConfigRefinedSchema } from './refined.js';
